Refactor all publisher views on its own module
